pyproject.toml
[tool.poetry]
name = "polyform-puzzle-solver"
version = "0.1.0"
description = "Solve Polyform Puzzle"
authors = ["kyunashige <kyuna.b23@gmail.com>"]
readme = "README.md"
packages = [{include = "polyform_puzzle_solver"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
numpy = "^1.24.2"
tqdm = "^4.65.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
